[refactor] Refactor App component - Created custom hook `useAuthState` to manage authentication state. - Extracted routing into a new AppRoutes component. - Updated `App.jsx` to use `useAuthState` and `AppRoutes`. - Improved code organization and readability for better maintainability and testability. - Fix duplicate key issue iin PhotoList component.
